+++ wp-content/plugins/core/src/Blocks/Blocks_Definer.php
@@ -8,6 +8,7 @@ use Tribe\Libs\Container\Definer_Interface;
 use Tribe\Project\Blocks\Types\Buttons\Buttons;
 use Tribe\Project\Blocks\Types\Card_Grid\Card_Grid;
 use Tribe\Project\Blocks\Types\Content_Columns\Content_Columns;
+use Tribe\Project\Blocks\Types\Content_Loop\Content_Loop;
 use Tribe\Project\Blocks\Types\Interstitial\Interstitial;
 use Tribe\Project\Blocks\Types\Accordion\Accordion;
 use Tribe\Project\Blocks\Types\Hero\Hero;
@@ -19,6 +20,7 @@ use Tribe\Project\Blocks\Types\Post_List\Post_List;
 use Tribe\Project\Blocks\Types\Quote\Quote;
 use Tribe\Project\Blocks\Types\Stats\Stats;
 use Tribe\Project\Blocks\Types\Tabs\Tabs;
+use Tribe\Project\Blocks\Types\Gallery_Grid\Gallery_Grid;
 
 class Blocks_Definer implements Definer_Interface {
 
@@ -30,10 +32,12 @@ class Blocks_Definer implements Definer_Interface {
 	public function define(): array {
 		return [
 			self::TYPES => DI\add( [
+				DI\get( Content_Loop::class ),
 				DI\get( Accordion::class ),
 				DI\get( Buttons::class ),
 				DI\get( Card_Grid::class ),
 				DI\get( Content_Columns::class ),
+				DI\get( Gallery_Grid::class ),
 				DI\get( Hero::class ),
 				DI\get( Media_Text::class ),
 				DI\get( Interstitial::class ),
@@ -59,6 +63,8 @@ class Blocks_Definer implements Definer_Interface {
 			 * Includes any 3rd-party block supported by this project, such as Gravity Forms.
 			 */
 			self::ALLOW_LIST => [
+				'acf/gallerygrid',
+				'acf/contentloop',
 				'acf/accordion',
 				'acf/buttons',
 				'acf/cardgrid',