+++ changelog.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## 2021.02
-* Fixed: An issue with the callback used in the `scroll-to` script. Also added automatic focus switching when using the `smooth-anchors` script.
+* Fixed: Add header and footer templates to ensure default TEC templates will load.
 
 ## 2021.01
 * Updated: selenium/standalone-chrome docker image to a tag that is updated frequently.